Update card to reuse withProductLightbox HOC.
