+++ client/a8c-for-agencies/sections/overview/sidebar/featured-woopayments/style.scss
@@ -45,6 +45,7 @@ $woocommerce-purple-50: #720EEC;
 	background-color: var(--color-surface);
 	color: $woocommerce-purple-50;
 	margin-block-end: 16px;
+	text-transform: uppercase;
 }
 
 .overview__featured-woopayments-logo {
